refactor: Money.percentageOf 반환 타입 BigDecimal + HALF_UP 반올림
- 소수점 둘째 자리 반올림(HALF_UP, scale=2)으로 정확한 할인율 표현
- 반환을 Int에서 BigDecimal로 전환하여 소수점 이하 정밀도 보존
- PointProductWithDiscount.discountRate, PointProductResponse.discountRate도 BigDecimal로 전파
- Vocabulary 설명과 예시값 갱신 (20 → 20.00)
- Money/PointProductWithDiscount 테스트 반영

+++ boot/ma-boot-web/src/main/kotlin/com/konkuk/ma/domain/point/api/response/PointProductResponse.kt
@@ -1,6 +1,7 @@
 package com.konkuk.ma.domain.point.api.response
 
 import com.konkuk.ma.domain.point.domain.PointProductWithDiscount
+import java.math.BigDecimal
 import java.time.LocalDate
 
 class PointProductResponse(
@@ -9,7 +10,7 @@ class PointProductResponse(
     val quantity: Int,
     val price: Int,
     val discountedPrice: Int,
-    val discountRate: Int,
+    val discountRate: BigDecimal,
     val discountType: String?,
     val isDiscountActive: Boolean,
 ) {
